requirements.txt
# Thai Law Chatbot — dependencies
openai>=1.40.0
python-dotenv>=1.0.0

# LangChain + vector store
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.1.0
chromadb>=0.5.0

# Embeddings model
sentence-transformers>=3.0.0

# FastAPI backend
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Data ingestion
tqdm
datasets>=2.20.0

# BM25
rank_bm25
